backend/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12.0
psycopg2-binary>=2.9.0
redis>=4.0.0
drf-spectacular>=0.21.0
django-encrypted-fields>=2.1.0
python-dotenv>=0.19.0
gunicorn>=20.1.0
pytest>=6.2.5
pytest-django>=4.4.0
django-cors-headers>=3.10.0
djangorestframework-simplejwt>=5.0.0
django-filter>=21.1 